Moved sub pokemons declarations to separate header file
